backend/requirements.txt
# Core dependencies
Flask==2.0.1
Flask-Cors==3.0.10
werkzeug==2.0.3
python-dotenv==0.21.0
pymongo==4.6.1

# AI & Machine Learning
google-generativeai>=0.5.0
tensorflow==2.13.0
tensorflow-hub==0.14.0

# Audio processing
numpy==1.24.3
scipy==1.10.1
sounddevice==0.4.6
SpeechRecognition==3.10.0
pyroomacoustics==0.6.0

# Optional dependencies for computer vision 
# Uncomment if needed for visual processing
opencv-python==4.7.0.72
Pillow==9.4.0
deepface==0.0.79

# Optional hardware dependencies
# Uncomment as needed for specific hardware integrations
# RPi.GPIO==0.7.1  # For Raspberry Pi GPIO
# adafruit-circuitpython-servokit==1.3.10  # For servo control
# rpi_ws281x==4.3.4  # For LED control

# Development & testing
pytest==7.4.0
pytest-flask==1.2.0

# Frontend dependencies are managed via package.json in the frontend directory 